Load web fonts through the fetcher, not the tab worker
