+++ crates/gosub_engine/src/engine/resource_pipeline/webfonts.rs
@@ -1,0 +1,308 @@
+//! `@font-face` web fonts: found in the document's stylesheets, fetched through the zone's
+//! fetcher, and registered with the font system before the document is handed to the tab.
+//!
+//! This used to run on the tab worker, synchronously, with a blocking fetch per font -- so a
+//! slow font server stopped the tab answering commands or drawing frames, and none of those
+//! requests went through the fetcher, which meant no policy, no cache and nothing in the
+//! network panel. Here the same work is a stage of the parse, where waiting costs a task
+//! rather than a tab.
+
+use crate::html::{EngineDocument, RenderConfiguration};
+use gosub_interface::css3::CssStylesheet as _;
+use gosub_interface::document::Document as _;
+use gosub_interface::font_system::FontSystem as _;
+use parking_lot::Mutex;
+use std::collections::HashSet;
+use std::sync::Arc;
+use url::Url;
+
+use super::html::{fetch_subresource, SubFetch};
+
+/// One `@font-face` to load: the CSS family it names, and the sources to try in order.
+struct Face {
+    family: String,
+    sources: Vec<Url>,
+}
+
+/// Fetch and register every web font the document's stylesheets declare.
+///
+/// Faces load concurrently; the sources within a face are tried in order and stop at the
+/// first that registers, because they are alternatives to each other (`src: url(a), url(b)`)
+/// rather than a list of things to fetch.
+pub(crate) async fn load_web_fonts<C: RenderConfiguration>(
+    doc: &EngineDocument<C>,
+    base_url: &Url,
+    font_system: &Arc<Mutex<C::FontSystem>>,
+    fetch: &SubFetch<'_>,
+    timing_scope: Option<gosub_shared::timing::ScopeId>,
+) {
+    let faces = collect_faces::<C>(doc, base_url);
+    if faces.is_empty() {
+        return;
+    }
+
+    let loaded = futures_util::future::join_all(faces.iter().map(|face| load_face(face, fetch, timing_scope))).await;
+
+    // Registration is serialised behind the font system's lock anyway, and doing it here --
+    // in the order the document declared the faces -- keeps the result independent of which
+    // fetch happened to finish first.
+    for (face, bytes) in faces.iter().zip(loaded) {
+        let Some((url, bytes)) = bytes else { continue };
+        let bytes = decode_web_font(bytes, &url);
+        match font_system.lock().register_font(bytes, Some(&face.family)) {
+            Ok(()) => log::debug!("Registered web font '{}' from {url}", face.family),
+            Err(e) => log::warn!("Failed to register web font '{}': {e:?}", face.family),
+        }
+    }
+}
+
+/// Every face worth loading, in document order, deduplicated by source URL.
+fn collect_faces<C: RenderConfiguration>(doc: &EngineDocument<C>, base_url: &Url) -> Vec<Face> {
+    let mut seen: HashSet<String> = HashSet::new();
+    let mut faces = Vec::new();
+
+    for sheet in doc.stylesheets() {
+        let sheet_url = Url::parse(sheet.url()).ok();
+        for (family, sources, unicode_range) in sheet.font_faces() {
+            // Google-style web fonts split a family into many `unicode-range` subsets (latin,
+            // cyrillic, greek, ...). We don't do per-glyph subset fallback, so register only
+            // subsets covering Basic Latin (and ranges with no descriptor), which covers
+            // Latin-script content without piling unusable subsets onto the same family.
+            if let Some(range) = &unicode_range {
+                if !unicode_range_covers_basic_latin(range) {
+                    continue;
+                }
+            }
+
+            let urls: Vec<Url> = sources
+                .iter()
+                .filter_map(|src| {
+                    let base = sheet_url.as_ref().unwrap_or(base_url);
+                    base.join(src).or_else(|_| base_url.join(src)).ok()
+                })
+                // A URL another face already claimed is that face's to register; a second
+                // family pointing at the same file would only fetch it twice.
+                .filter(|url| seen.insert(url.to_string()))
+                .collect();
+
+            if !urls.is_empty() {
+                faces.push(Face { family, sources: urls });
+            }
+        }
+    }
+    faces
+}
+
+/// Try a face's sources in order and return the first that produced bytes.
+async fn load_face(
+    face: &Face,
+    fetch: &SubFetch<'_>,
+    timing_scope: Option<gosub_shared::timing::ScopeId>,
+) -> Option<(Url, Vec<u8>)> {
+    for url in &face.sources {
+        let started = std::time::Instant::now();
+        let body = fetch_subresource(url.as_str(), crate::net::types::ResourceKind::Font, fetch).await;
+        let elapsed = started.elapsed().as_micros() as u64;
+
+        // Recorded against the navigation explicitly: this runs across awaits, where a
+        // thread-local scope does not hold.
+        match timing_scope {
+            Some(scope) => gosub_shared::timing::record_in(scope, "net.fetch.font", elapsed, Some(url.to_string())),
+            None => gosub_shared::timing::record("net.fetch.font", elapsed, Some(url.to_string())),
+        }
+
+        match body {
+            Some((_, bytes)) if !bytes.is_empty() => return Some((url.clone(), bytes)),
+            _ => log::warn!("Web font fetch {url} produced nothing"),
+        }
+    }
+    None
+}
+/// Whether a CSS `unicode-range` descriptor (e.g. `"U+0000-00FF, U+0131"`) includes the
+/// Basic-Latin letter `U+0041` ('A') - our proxy for "covers Latin-script text".
+fn unicode_range_covers_basic_latin(range: &str) -> bool {
+    const TARGET: u32 = 0x41; // 'A'
+    for token in range.split([',', ' ', '\t', '\n', '\r']).filter(|t| !t.is_empty()) {
+        let Some(hex) = token
+            .trim()
+            .strip_prefix("U+")
+            .or_else(|| token.trim().strip_prefix("u+"))
+        else {
+            continue;
+        };
+        let (lo, hi) = match hex.split_once('-') {
+            Some((a, b)) => (parse_hex_bound(a, false), parse_hex_bound(b, true)),
+            None => (parse_hex_bound(hex, false), parse_hex_bound(hex, true)),
+        };
+        if let (Some(lo), Some(hi)) = (lo, hi) {
+            if lo <= TARGET && TARGET <= hi {
+                return true;
+            }
+        }
+    }
+    false
+}
+
+/// Unwrap a downloaded web-font payload into raw SFNT bytes the font backends can decode.
+///
+/// WOFF2 (magic `wOF2`) is a Brotli-compressed wrapper around an OpenType/TrueType font,
+/// with the `glyf`/`loca` tables stored in a transformed form. Skia and fontconfig don't
+/// decode it (e.g. Google Fonts serves WOFF2 to modern UAs like ours), so we decompress it
+/// to a flat SFNT here. Bare SFNT (`OTTO`/`true`/`ttcf`/`0x00010000`) and anything we don't
+/// recognise are returned unchanged - including WOFF1, which the backends already handle.
+/// On a decode error we log and return the original bytes so the subsequent `register_font`
+/// surfaces a single, consistent failure path.
+fn decode_web_font(bytes: Vec<u8>, font_url: &Url) -> Vec<u8> {
+    const WOFF2_MAGIC: &[u8; 4] = b"wOF2";
+    if bytes.len() < 4 || &bytes[0..4] != WOFF2_MAGIC {
+        return bytes;
+    }
+    match woff2_to_sfnt(&bytes) {
+        Ok(sfnt) => {
+            log::debug!(
+                "Decoded WOFF2 web font from {font_url} ({} → {} bytes)",
+                bytes.len(),
+                sfnt.len()
+            );
+            sfnt
+        }
+        Err(e) => {
+            log::warn!("Failed to decode WOFF2 web font from {font_url}: {e}");
+            bytes
+        }
+    }
+}
+
+/// Decompress a WOFF2 font to a flat SFNT (TTF/OTF) byte buffer. allsorts handles the Brotli
+/// decompression and the `glyf`/`loca` transform reconstruction; we then re-assemble the
+/// reconstructed tables into the on-disk SFNT layout (offset table + table directory + 4-byte
+/// aligned table data) that font backends expect.
+fn woff2_to_sfnt(bytes: &[u8]) -> Result<Vec<u8>, String> {
+    use allsorts::binary::read::ReadScope;
+    use allsorts::woff2::Woff2Font;
+
+    let font = ReadScope::new(bytes)
+        .read::<Woff2Font<'_>>()
+        .map_err(|e| format!("parse: {e:?}"))?;
+    let sfnt_version = font.flavor();
+    let tables = font
+        .table_provider(0)
+        .map_err(|e| format!("reconstruct: {e:?}"))?
+        .into_tables();
+
+    Ok(assemble_sfnt(sfnt_version, tables))
+}
+
+/// Pack a set of font tables into an SFNT byte buffer per the OpenType spec: a 12-byte offset
+/// table, a 16-byte directory entry per table (sorted by tag), then each table's data padded to
+/// a 4-byte boundary. Per-table checksums are computed; the `head` table's `checkSumAdjustment`
+/// is left as-is (font backends parse without validating it).
+fn assemble_sfnt(sfnt_version: u32, tables: std::collections::HashMap<u32, Box<[u8]>>) -> Vec<u8> {
+    let mut entries: Vec<(u32, Box<[u8]>)> = tables.into_iter().collect();
+    entries.sort_by_key(|(tag, _)| *tag);
+    let num_tables = entries.len() as u16;
+
+    // Binary-search hint fields: largest power of two <= num_tables.
+    let mut entry_selector = 0u16;
+    while (1u16 << (entry_selector + 1)) <= num_tables {
+        entry_selector += 1;
+    }
+    let search_range = (1u16 << entry_selector) * 16;
+    let range_shift = num_tables.wrapping_mul(16).wrapping_sub(search_range);
+
+    let mut directory = Vec::with_capacity(16 * entries.len());
+    let mut data = Vec::new();
+    let mut offset = 12 + 16 * entries.len();
+    for (tag, table) in &entries {
+        directory.extend_from_slice(&tag.to_be_bytes());
+        directory.extend_from_slice(&sfnt_table_checksum(table).to_be_bytes());
+        directory.extend_from_slice(&(offset as u32).to_be_bytes());
+        directory.extend_from_slice(&(table.len() as u32).to_be_bytes());
+        data.extend_from_slice(table);
+        while data.len() % 4 != 0 {
+            data.push(0);
+        }
+        offset += (table.len() + 3) & !3;
+    }
+
+    let mut out = Vec::with_capacity(12 + directory.len() + data.len());
+    out.extend_from_slice(&sfnt_version.to_be_bytes());
+    out.extend_from_slice(&num_tables.to_be_bytes());
+    out.extend_from_slice(&search_range.to_be_bytes());
+    out.extend_from_slice(&entry_selector.to_be_bytes());
+    out.extend_from_slice(&range_shift.to_be_bytes());
+    out.extend_from_slice(&directory);
+    out.extend_from_slice(&data);
+    out
+}
+
+/// SFNT table checksum: the sum of the table's contents read as big-endian `u32`s, with the
+/// final partial word zero-padded, in wrapping (mod 2^32) arithmetic.
+fn sfnt_table_checksum(data: &[u8]) -> u32 {
+    let mut sum = 0u32;
+    for chunk in data.chunks(4) {
+        let mut word = [0u8; 4];
+        word[..chunk.len()].copy_from_slice(chunk);
+        sum = sum.wrapping_add(u32::from_be_bytes(word));
+    }
+    sum
+}
+
+/// Parse a `unicode-range` hex bound, expanding `?` wildcards to `0` (low bound) or `F`
+/// (high bound), e.g. `U+00??` → `0x0000..=0x00FF`.
+fn parse_hex_bound(s: &str, high: bool) -> Option<u32> {
+    let s = s.trim();
+    if s.is_empty() {
+        return None;
+    }
+    let filled: String = s
+        .chars()
+        .map(|c| {
+            if c == '?' {
+                if high {
+                    'F'
+                } else {
+                    '0'
+                }
+            } else {
+                c
+            }
+        })
+        .collect();
+    u32::from_str_radix(&filled, 16).ok()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Verify `decode_web_font` turns a real WOFF2 payload into an SFNT the font stack can
+    /// parse. Reads the fixture path from `GOSUB_WOFF2_FIXTURE` so we neither hit the network
+    /// nor commit a binary font; skips when unset.
+    #[test]
+    fn decode_web_font_woff2_roundtrips_to_sfnt() {
+        let Ok(path) = std::env::var("GOSUB_WOFF2_FIXTURE") else {
+            eprintln!("skipping: set GOSUB_WOFF2_FIXTURE to a .woff2 file to run");
+            return;
+        };
+        let woff2 = std::fs::read(&path).expect("read fixture");
+        assert_eq!(&woff2[0..4], b"wOF2", "fixture must be WOFF2");
+
+        let url = url::Url::parse("https://example.test/font.woff2").unwrap();
+        let sfnt = decode_web_font(woff2, &url);
+
+        // Output must be a different, valid SFNT (TrueType `0x00010000` or OpenType `OTTO`).
+        let magic = u32::from_be_bytes([sfnt[0], sfnt[1], sfnt[2], sfnt[3]]);
+        assert!(magic == 0x0001_0000 || magic == 0x4F54_544F, "not SFNT: {magic:#010x}");
+
+        // It must re-parse and expose the core tables a backend reads.
+        use allsorts::binary::read::ReadScope;
+        use allsorts::font_data::FontData;
+        use allsorts::tables::FontTableProvider;
+        let font = ReadScope::new(&sfnt).read::<FontData<'_>>().expect("parse SFNT");
+        let provider = font.table_provider(0).expect("table provider");
+        for tag in [allsorts::tag::HEAD, allsorts::tag::CMAP, allsorts::tag::GLYF] {
+            assert!(provider.has_table(tag), "missing table {tag:#010x}");
+        }
+    }
+}
